v2/dash/requirements.txt
psycopg2-binary
Dash
Pandas
dash-bootstrap-components
yfinance
pandas_datareader
sqlalchemy